+++ pyclient/cookiejar.py
@@ -1,19 +1,3 @@
-#!/usr/bin/env python3
-
-# Copyright 2018 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ------------------------------------------------------------------------------
 '''
 Command line interface for cookiejar TF.
 Parses command line arguments and passes to the CookieJarClient class
@@ -27,9 +11,9 @@ import sys
 import traceback
 
 from colorlog import ColoredFormatter
-from cookiejar_client import CookieJarClient
+from notary_client import NotaryClient
 
-KEY_NAME = 'mycookiejar'
+KEY_NAME = 'mynotary'
 
 # hard-coded for simplicity (otherwise get the URL from the args in main):
 #DEFAULT_URL = 'http://localhost:8008'